renamed events file extension
